Typescriptify Users-Component
